+++ CHANGELOG.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## Changed
 - Commit messages now use "🔗" emoji directly instead of ":link:"
-- Streamlined logging messages
+- Streamlined and improved logging messages; switched to Signale
 
 
 # 0.9.5 - 2021-01-16